pyproject.toml
[project]
name = "aishwarya-ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "openai>=1.76.0",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
]
